+++ test/unit/src/pipelineSteps/applicationFramework/vue.spec.ts
@@ -113,7 +113,7 @@ describe("Vue", () => {
             Chai.expect(engineVariablesStub.getConfiguration<JestConfiguration>("Jest").moduleNameMapper["\\.vue$"]).to.be.equal("<rootDir>/test/unit/dummy.mock.js");
             Chai.expect(engineVariablesStub.buildTranspileInclude.length).to.be.equal(3);
             Chai.expect(engineVariablesStub.buildTranspilePreBuild.length).to.be.equal(1);
-            Chai.expect(engineVariablesStub.buildTranspilePostBuild.length).to.be.equal(1);
+            Chai.expect(engineVariablesStub.buildTranspilePostBuild.length).to.be.equal(4);
         });
 
         it("can be called with configurations as not requirejs", async () => {
@@ -136,7 +136,7 @@ describe("Vue", () => {
             Chai.expect(engineVariablesStub.getConfiguration<JestConfiguration>("Jest").moduleNameMapper["\\.vue$"]).to.be.equal("<rootDir>/test/unit/dummy.mock.js");
             Chai.expect(engineVariablesStub.buildTranspileInclude.length).to.be.equal(1);
             Chai.expect(engineVariablesStub.buildTranspilePreBuild.length).to.be.equal(0);
-            Chai.expect(engineVariablesStub.buildTranspilePostBuild.length).to.be.equal(1);
+            Chai.expect(engineVariablesStub.buildTranspilePostBuild.length).to.be.equal(4);
         });
 
         it("can be called no configurations with false mainCondition", async () => {